facto setupProject
