+++ plugins/scarb/__test__/scarb/install.spec.ts
@@ -2,7 +2,7 @@
 import { describe, it, expect, beforeEach, jest } from '@jest/globals';
 import { installScarb } from '../../src/actions/installScarb.js';
 import { createMockStarknetAgent } from '../jest/setEnvVars.js';
-import * as environmentModule from '../../src/utils/environment.js';
+import * as environmentModule from '../../src/utils/install.js';
 import { exec } from 'child_process';
 import { promisify } from 'util';
