+++ plugins/scarb/src/actions/buildContract.ts
@@ -1,11 +1,13 @@
 import { StarknetAgentInterface } from '@starknet-agent-kit/agents';
-import * as fs from 'fs/promises';
-import * as path from 'path';
-import { initProject, buildProject,  isProjectInitialized, cleanProject } from '../utils/project.js';
-import { addSeveralDependancies, importContract, cleanLibCairo, checkWorkspaceLimit, getGeneratedContractFiles } from '../utils/index.js';
-import { getWorkspacePath, resolveContractPath } from '../utils/path.js';
-import { checkScarbInstalled, getScarbInstallInstructions } from '../utils/environment.js';
-import { addTomlSection } from '../utils/toml-utils.js';
+import { buildProject } from '../utils/command.js';
+import { setupScarbProject } from '../utils/common.js';
+import { 
+  addSeveralDependancies, 
+  importContract, 
+  cleanLibCairo,  
+  addTomlSection, 
+  getGeneratedContractFiles
+} from '../utils/preparation.js';
 
 export interface Dependency {
   name: string;
@@ -25,30 +27,12 @@ export const compileContract = async (
   params: CompileContractParams
 ) => {
   try {
-    const isScarbInstalled = await checkScarbInstalled();
-    if (!isScarbInstalled) {
-      return JSON.stringify({
-        status: 'failure',
-        error: await getScarbInstallInstructions(),
-      });
-    }
+    const { projectDir, resolvedPaths } = await setupScarbProject({
+      projectName: params.projectName,
+      filePaths: params.contractPaths,
+      dependencies: params.dependencies
+    });
     
-    const workspaceDir = getWorkspacePath();
-    try {
-      await fs.mkdir(workspaceDir, { recursive: true });
-    } catch (error) {}
-
-    const projectDir = path.join(workspaceDir, params.projectName);
-    const contractPaths = params.contractPaths.map(contractPath => 
-      resolveContractPath(contractPath)
-    );
-
-    await checkWorkspaceLimit(workspaceDir, params.projectName);
-    const isInitialized = await isProjectInitialized(projectDir);
-    if (!isInitialized) {
-      await initProject(agent, { name: params.projectName, projectDir });
-    }
-
     await addTomlSection({
       workingDir: projectDir,
       sectionTitle: 'target.starknet-contract',
@@ -60,12 +44,12 @@ export const compileContract = async (
 
     await addSeveralDependancies(params.dependencies || [], projectDir);
     await cleanLibCairo(projectDir);
-    for (const contractPath of contractPaths) {
+    for (const contractPath of resolvedPaths) {
       await importContract(contractPath, projectDir);
     }
 
     // await cleanProject(agent, { path: projectDir });
-    const buildResult = await buildProject(agent, { path: projectDir });
+    const buildResult = await buildProject({ path: projectDir });
     const parsedBuildResult = JSON.parse(buildResult);
 
     const contractFiles = await getGeneratedContractFiles(projectDir);
@@ -80,8 +64,8 @@ export const compileContract = async (
       casmFiles: contractFiles.casmFiles,
       projectDir: projectDir
     });
-    
   } catch (error) {
+    console.log("Error compiling contract:", error);
     return JSON.stringify({
       status: 'failure',
       error: error instanceof Error ? error.message : 'Unknown error',
