+++ plugins/scarb/src/utils/install.ts
